requirement_dev.txt
appdirs==1.4.4
black==20.8b1
boto3==1.14.52
botocore==1.17.52
click==7.1.2
docutils==0.15.2
fakefiles==0.1
jmespath==0.10.0
mimesis==4.1.2
mypy-extensions==0.4.3
pathspec==0.8.0
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
